PP-14542 - rename model, change where referenced, better redirect_result description
