+++ openapi/connector_spec.yaml
@@ -3551,7 +3551,7 @@ components:
           type: string
         redirect_result:
           type: string
-          description: Adyen 3DS2 URL-encoded Redirect Result data
+          description: Adyen 3DS2 Redirect Result data
     AuthCardDetails:
       type: object
       properties:
