+++ src/main/java/uk/gov/pay/connector/gateway/adyen/request/Adyen3dsAuthorisationRequest.java
@@ -1,23 +1,23 @@
 package uk.gov.pay.connector.gateway.adyen.request;
 
-import jakarta.ws.rs.core.MediaType;
 import uk.gov.pay.connector.gateway.GatewayOrder;
 import uk.gov.pay.connector.gateway.PaymentGatewayName;
-import uk.gov.pay.connector.gateway.adyen.request.json.PaymentDetailsRequestPayload;
-import uk.gov.pay.connector.gateway.model.OrderRequestType;
+import uk.gov.pay.connector.gateway.adyen.request.json.Authorise3dsRequestPayload;
 import uk.gov.pay.connector.gateway.model.request.GatewayClientPostRequest;
 import uk.gov.pay.connector.util.JsonObjectMapper;
 
 import java.net.URI;
 import java.util.Map;
 
+import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON_TYPE;
 import static uk.gov.pay.connector.gateway.PaymentGatewayName.ADYEN;
+import static uk.gov.pay.connector.gateway.model.OrderRequestType.AUTHORISE_3DS;
 
 public record Adyen3dsAuthorisationRequest(
         URI url,
         Map<String, String> headers,
         String gatewayAccountType,
-        PaymentDetailsRequestPayload requestPayload,
+        Authorise3dsRequestPayload requestPayload,
         JsonObjectMapper jsonObjectMapper
 ) implements GatewayClientPostRequest {
     @Override
