+++ src/main/java/uk/gov/pay/connector/gateway/model/Auth3dsResult.java
@@ -25,7 +25,7 @@ public class Auth3dsResult {
     @JsonIgnore
     private String gatewayResponseStringified;
 
-    @Schema(description = "Adyen 3DS2 URL-encoded Redirect Result data")
+    @Schema(description = "Adyen 3DS2 Redirect Result data")
     public String getRedirectResult() {
         return redirectResult;
     }
